feat: Task2 added

+++ src/Task2/array_sum/array_sum.cpp
@@ -1,7 +1,7 @@
 #include "array_sum.hpp"
 
-[[nodiscard]] int arraySum(const int* arr, const std::size_t size) {
-    int result = 0;
+[[nodiscard]] int64_t arraySum(const int* arr, const std::size_t size) {
+    int64_t result = 0;
     for (std::size_t i = 0; i < size; ++i) {
         result += arr[i];
     }
