+++ src/Task2/array_sum/array_sum.hpp
@@ -2,7 +2,8 @@
 #define HIGH_PERF_SOFTWARE_CS_SRC_TASK2_ARRAY_SUM_ARRAY_SUM_HPP_
 
 #include <cstddef>
+#include <cstdint>
 
-[[nodiscard]] int arraySum(const int* arr, const std::size_t size);
+[[nodiscard]] int64_t arraySum(const int* arr, const std::size_t size);
 
 #endif  // HIGH_PERF_SOFTWARE_CS_SRC_TASK2_ARRAY_SUM_ARRAY_SUM_HPP_
